add securiter and fix inscription

+++ app/src/Controller/DevisController.php
@@ -17,6 +17,7 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/devis')]
 class DevisController extends AbstractController
 {
+    
     private $knpSnappyPdf;
 
     public function __construct(Pdf $knpSnappyPdf)
@@ -27,6 +28,10 @@ class DevisController extends AbstractController
     #[Route('/', name: 'devis_index', methods: ['GET'])]
     public function index(DevisRepository $devisRepository): Response
     {
+        //check permittion 
+        if (!$this->getUser()) {
+                 return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         return $this->render('devis/index.html.twig', [
             'devis' => $devisRepository->findAll(),
         ]);
@@ -35,6 +40,10 @@ class DevisController extends AbstractController
     #[Route('/new', name: 'devis_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+            //check permittion 
+        if (!$this->getUser()) {
+                return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         $devis = new Devis();
         $form = $this->createForm(DevisType::class, $devis);
         $form->handleRequest($request);
@@ -61,6 +70,10 @@ class DevisController extends AbstractController
     #[Route('/{id}', name: 'devis_show', methods: ['GET'])]
     public function show(Devis $devis): Response
     {
+            //check permittion 
+            if (!$this->getUser()) {
+                return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         return $this->render('devis/show.html.twig', [
             'devis' => $devis,
         ]);
@@ -69,6 +82,10 @@ class DevisController extends AbstractController
     #[Route('/{id}/edit', name: 'devis_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Devis $devis, EntityManagerInterface $entityManager): Response
     {
+        //check permittion 
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         $form = $this->createForm(DevisType::class, $devis);
         $form->handleRequest($request);
 
@@ -87,6 +104,10 @@ class DevisController extends AbstractController
     #[Route('/{id}', name: 'devis_delete', methods: ['POST'])]
     public function delete(Request $request, Devis $devis, EntityManagerInterface $entityManager): Response
     {
+             //check permittion 
+        if (!$this->getUser()) {
+                return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         if ($this->isCsrfTokenValid('delete'.$devis->getId(), $request->request->get('_token'))) {
             $entityManager->remove($devis);
             $entityManager->flush();
@@ -98,6 +119,10 @@ class DevisController extends AbstractController
     #[Route('/{id}/pdf', name: 'devis_pdf', methods: ['GET'])]
     public function pdf(Devis $devis): Response
     {
+            //check permittion 
+        if (!$this->getUser()) {
+                return $this->redirectToRoute('app_login'); // Redirect to login page if not authenticated
+        }
         $html = $this->renderView('devis/pdf.html.twig', [
             'devis' => $devis,
         ]);
